backend/requirements.txt
# FastAPI & Web
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sse-starlette>=1.8.0

# LLM
openai>=1.12.0

# 数据验证
pydantic>=2.0.0

# Obsidian
python-frontmatter>=1.1.0

# 配置
python-dotenv>=1.0.0

# 工具
aiofiles>=23.0.0

# 打包
pyinstaller>=6.0.0